requirements.txt
langchain
openai
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
langchain_community
faiss-cpu
streamlit
langchain-groq
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
bs4
langchain_chroma
streamlit
pymupdf
pypdf
arxiv
wikipedia
duckduckgo-search
transformers
langchain-huggingface
huggingface_hub
duckdb
validators == 0.28.1
unstructured
sentence_transformers
numexpr
openpyxl
langgraph
datasets